requirements.txt
spotipy~=2.16.1
pandas~=1.2.0
numpy~=1.19.4
Keras~=2.4.3
opencv-python~=4.4.0.46
scikit-learn~=0.24.0
config~=0.5.0.post0
Pillow~=8.1.0
sklearn~=0.0